fix: improve block header layout with responsive design
- Move CopyLinkButton to shared component for reuse
- Consolidate block variant header into VariantSection component
- Title + version + link button on left, npx command on right
- Use flex-wrap to move npx command below title on mobile to prevent overflow
- Clean up unused imports from page.tsx
